requirements.txt
absl-py==1.4.0
altair==4.2.2
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
decorator==5.1.1
Django==4.2
docutils==0.19
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.95.1
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
image==1.5.33
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lightgbm==3.3.5
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pip==23.1.2
pluggy==1.0.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==58.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
statistics==1.0.3.5
streamlit==1.21.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.0
wheel==0.40.0
wrapt==1.14.1
xgboost==1.7.5
zipp==3.15.0
zope.interface==6.0
nltk
sastrawi
gensim
